refactor: move Backup sync functionality to backup.py
